// src/types/permissions.ts
export type PermissionKey =
  | 'canReadDashboard'
  | 'canScanId'
  | 'canReadStudents'
  | 'canWriteStudents'
  | 'canCreateStudents'
  | 'canDeleteStudents'
  | 'canExportStudents'
  | 'canReadAttendance'
  | 'canExportAttendance'
  | 'canReadActivityLog'
  | 'canReadUsers'
  | 'canWriteUsers'
  | 'canManageSiteSettings'
  | 'canSeeAllRecords';